Added pagination for both tune and user lists
This is belt-and-braces at the moment.  Pagination information  is
returned both in the JSON and in the custom musicrest-pagination
response header.  I'll see if we can remove the latter if the frontend
might support it.

We take account of the configured pagination page size as a default but
allow this to be over-ridden by the client.  Both page and size query
params are supported.