Use a date string not ts in TuneRefs
It seems very awkward in postgresql to return a string representing an
epoch time from a timestamp field.  It is much, much simpler to reformat
the timestamp as an appropriately formatted date and just return that in
the JSON.  I suppose we lose the ability to support different clients
(in different languages) but I'm not thinking of doing that anyway.