Normalise rhythm and key signature in searches
Use the normalisation functions from the abc-header-parser for query
parameters of these types.  Normalisation now occurs both during
validation of a new ABC tune and also when requesting that parameter
when querying the tune.  This is an attempt to regularise these fields
from the morass of different spellings out in the wild.