Add AbcMetadataSubmission
We use one form of Metadata to write to the DB (because the DB populates
the creation timestamp and synthetic id).  We use another form for
reading from the database (which must, for example, reead the timestamp).

We place FromRow, ToRow accordingly.

The buildMetadata function is for use only with the former, but we seem
to have to continue with a function buildMetadata for reading simply to
satisfy the testing client, so it is now called buildClientMetadata.