denormalise tunes
It used to have a foreign reference to user_id to identify the submitter
of the tune, now it denormalises just to have the user name.  I agonized
about whether or not to make this change but I think it's the right one.

We want the name of the submitter to be attached to the tune, but if
that user de-registers we don't want to lose the tune (or the name).
Thus we don't need over-complex on delete referential intergrity rules.  Also, the implementation is simplified - we already have checked the credentials of the user if a tune is to be submitted and we don't need to go back to the database to get hold of the id

The same considerations apply to comments.