Improved the unit tests
Firstly, a lot of tests were failing in a random manner with connection
issues.  I've tried to correct this by putting in a delay of about 0.1
or 0.2 seconds in the bracketing of the tests to allow the server time
to start up.  This seems to improve things a bit.

This aside, got the tune tests working.